Solution to Input and Print in Hexadecimal and Octal

+++ Input and Print in Hexadecimal and Octal@@6788d5012db25efec946f941/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main()
 {
-    float a;
-    scanf("%f",&a);
+    int a;
+    scanf("%d",&a);
     printf("Hexadecimal: %x\n",a);
     printf("Octal: %o",a);
     return 0;
